requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
Django==2.2.5
django-bootstrap4==2.3.1
django-phone-field==1.8.1
djangorestframework==3.12.1
psycopg2-binary==2.8.6
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.4.1
